refactor(data): migrate various classes imports to reorganized exceptions and models
